fix bug in delete_report

+++ app/reports/views.py
@@ -189,20 +189,22 @@ def delete_report(report_id):
     """Delete a report"""
 
     report = IncidentReport.query.filter_by(id=report_id).first()
+    report_user_id = None
 
-    if report.picture_deletehash:
-        # Asynchronously delete the report's image
-        get_queue().enqueue(
-            delete_image,
-            deletehash=report.picture_deletehash,
-            imgur_client_id=current_app.config['IMGUR_CLIENT_ID'],
-            imgur_client_secret=current_app.config['IMGUR_CLIENT_SECRET'],
-        )
-    report_user_id = report.user_id
+    if report is not None:
+        if report.picture_deletehash:
+            # Asynchronously delete the report's image
+            get_queue().enqueue(
+                delete_image,
+                deletehash=report.picture_deletehash,
+                imgur_client_id=current_app.config['IMGUR_CLIENT_ID'],
+                imgur_client_secret=current_app.config['IMGUR_CLIENT_SECRET'],
+            )
+        report_user_id = report.user_id
 
-    db.session.delete(report)
-    db.session.commit()
-    flash('Successfully deleted report.', 'success')
+        db.session.delete(report)
+        db.session.commit()
+        flash('Successfully deleted report.', 'success')
 
     # TODO - address edge case where an admin clicks on their own report from
     # reports/all endpoint, should redirect back to /all. use cookies
